Add react-router dependancy to front end
Added React Router to front end to route between pages. Broke router into separate component to remove from rest of the component hierarchy.

Needed to hard code the front end url to the server redirect after passport authentication. Not sure if this needs to be linked up with a .env variable.