Add theme, spotify, and firebase contexts to client
Setup the initial context providers to access data globally on the front end. Decided to keep the spotify and firebase contexts separate from the theme context to keep things a bit cleaner and to separate concerns of each.

Also imported styled-components to inject the theme provider throughout the app via the router component. See comments for more details