Add tuneroom route and break out api functionality
Added a 'tuneroom' route that will house the search functionality and also render the playlist. polished the UX a bit to close the search modal and to route logically between routes.

ToDo: adding a song to the playlist doesn't update it in real time. I'm wondering if this should be integrated with the SpotifyContext to render this on the front end.