Add realtime updates to tuneroom component
User can search for a song in the tuneroom, and add that song. Song will populate the firestore, then get rendered via context throughout the app. Also modified the firestore to add a 'votes' key into the initial song creation to address NaN bug for voting for a newly created song.