Mutate state in playlist creation
On playlist creation, the various playlist properties are set on creation, and then context sets them based on localStorage to persist through page refresh.

There is still a bug. When on a playlist, and user clicks 'back' then rejoins a room, the playlist does not autoplay. Interestingly, when you manually add the /join parameter to the url string, then navigate to a room, it DOES autoplay...something to do with browser history maybe?